Move make_public_only_session into its own utils package
