+++ corehq/motech/auth.py
@@ -8,13 +8,12 @@ import attr
 import requests
 from oauthlib.oauth2 import LegacyApplicationClient, BackendApplicationClient
 from requests import Session
-from requests.adapters import HTTPAdapter
 from requests.auth import AuthBase, HTTPBasicAuth, HTTPDigestAuth
 from requests.exceptions import RequestException
 from requests_oauthlib import OAuth1, OAuth2Session
 
 from corehq.motech.exceptions import ConfigurationError
-from corehq.motech.utils import get_endpoint_url
+from corehq.util.public_only_requests.public_only_requests import make_session_public_only
 
 if TYPE_CHECKING:
     from corehq.motech.models import ConnectionSettings
@@ -82,31 +81,6 @@ class HTTPBearerAuth(AuthBase):
         token = self._get_auth_token(r)
         r.headers["Authorization"] = f"Bearer {token}"
         return r
-
-
-class PublicOnlyHttpAdapter(HTTPAdapter):
-    def __init__(self, domain_name, src):
-        self.domain_name = domain_name
-        self.src = src
-        super().__init__()
-
-    def get_connection(self, url, proxies=None):
-        from corehq.motech.requests import validate_user_input_url_for_repeaters
-        validate_user_input_url_for_repeaters(url, domain=self.domain_name, src=self.src)
-        return super().get_connection(url, proxies=proxies)
-
-
-def make_session_public_only(session, domain_name, src):
-    """
-    Modifies `session` to validate urls before sending and accept only hosts resolving to public IPs
-
-    Once this function has been called on a session, session.request, etc., will
-    raise PossibleSSRFAttempt whenever called with a url host that resolves to a non-public IP.
-    """
-    # the following two lines entirely replace the default adapters with our custom ones
-    # by redefining the adapter to use for the two default prefixes
-    session.mount('http://', PublicOnlyHttpAdapter(domain_name=domain_name, src=src))
-    session.mount('https://', PublicOnlyHttpAdapter(domain_name=domain_name, src=src))
 
 
 class AuthManager:
