+++ corehq/util/public_only_requests/public_only_requests.py
@@ -1,0 +1,26 @@
+from requests.adapters import HTTPAdapter
+
+
+def make_session_public_only(session, domain_name, src):
+    """
+    Modifies `session` to validate urls before sending and accept only hosts resolving to public IPs
+
+    Once this function has been called on a session, session.request, etc., will
+    raise PossibleSSRFAttempt whenever called with a url host that resolves to a non-public IP.
+    """
+    # the following two lines entirely replace the default adapters with our custom ones
+    # by redefining the adapter to use for the two default prefixes
+    session.mount('http://', PublicOnlyHttpAdapter(domain_name=domain_name, src=src))
+    session.mount('https://', PublicOnlyHttpAdapter(domain_name=domain_name, src=src))
+
+
+class PublicOnlyHttpAdapter(HTTPAdapter):
+    def __init__(self, domain_name, src):
+        self.domain_name = domain_name
+        self.src = src
+        super().__init__()
+
+    def get_connection(self, url, proxies=None):
+        from corehq.motech.requests import validate_user_input_url_for_repeaters
+        validate_user_input_url_for_repeaters(url, domain=self.domain_name, src=self.src)
+        return super().get_connection(url, proxies=proxies)
